Milestone: Amiga can now perform flash id
Added "flash" memory space for STM32 internal flash to "d" command
Amiga flash config, including bank info is now saved to STM32 flash
KS ROM bank can be automatically set at cold power-on and Amiga reset
Added "reset amiga long" command option
Now automatically detects 16-bit or 32-bit Amiga mode
Added AmigaOS "smash" command
Implemented message interface between AmigaOS and Kicksmash

+++ fw/prom_access.h
@@ -17,7 +17,7 @@ rc_t prom_write(uint32_t addr, uint width, void *bufp);
 rc_t prom_erase(uint mode, uint32_t addr, uint32_t len);
 rc_t prom_read_binary(uint32_t addr, uint32_t len);
 rc_t prom_write_binary(uint32_t addr, uint32_t len);
-void prom_cmd(uint32_t addr, uint16_t cmd);
+void prom_cmd(uint32_t addr, uint32_t cmd);
 rc_t prom_id(void);
 void prom_disable(void);
 rc_t prom_status(void);
